Rename 'services' layer to 'application' layer
